Improved code quality, null safety, and input handling

+++ Data/ProductRepository.cs
@@ -8,14 +8,14 @@ namespace LegacyOrderService.Data
 {
     public class ProductRepository : IProductRepository
     {
-        private readonly Dictionary<string, double> _productPrices = new()
+        private readonly Dictionary<string, decimal> _productPrices = new()
         {
-            ["Widget"] = 12.99,
-            ["Gadget"] = 15.49,
-            ["Doohickey"] = 8.75
+            ["Widget"] = 12.99M,
+            ["Gadget"] = 15.49M,
+            ["Doohickey"] = 8.75M
         };
 
-        public double GetPrice(string productName)
+        public decimal GetPrice(string productName)
         {
             // Simulate an expensive lookup
             Thread.Sleep(500);
